+++ CapaModelo/DetalleVenta.cs
@@ -10,8 +10,10 @@ namespace CapaModelo
     {
         public int Cantidad { get; set; }
         public string NombreProducto { get; set; }
-        public float PrecioUnidad { get; set; }
-        public float ImporteTotal { get; set; }
-        public float ImporteTotalIvaIncluido { get; set; }
+        // FIX: Cambiado de float a decimal para evitar errores de precisión en montos monetarios.
+        // float tiene ~7 dígitos de precisión, decimal(18,2) es el estándar para dinero.
+        public decimal PrecioUnidad { get; set; }
+        public decimal ImporteTotal { get; set; }
+        public decimal ImporteTotalIvaIncluido { get; set; }
     }
 }
